Split SQL corrector dump files into a real Laravel tree
sey.php と "SqlSyntaxCorrectorService .php" は、6 ファイル分のコードを
1 ファイルに連結したものだった。

- sey.php は同じ内容を 2 回繰り返しており、833 行目の `*/<?php` で
  Parse error になっていた（PHP から一切ロードできない状態）。
- "SqlSyntaxCorrectorService .php" はファイル名に半角スペースを含み、
  PSR-4 オートロードにも CLI 引数にも扱いにくい。
- 両者の中身は同一（sey.php 側にセットアップ手順コメントが付くのみ）。

Laravel/sql-corrector/ 以下に、Laravel の実際の配置へ分割:

  app/Services/SqlSyntaxCorrectorService.php
  app/Http/Controllers/SqlCorrectorController.php
  app/Console/Commands/SqlCorrectorCommand.php
  app/Providers/SqlCorrectorServiceProvider.php
  routes/api.php
  resources/views/sql-corrector.blade.php
  tests/SqlCorrectorExample.php
  README.md  (コメントで書かれていたセットアップ手順を移設)

コードの中身は変更していない。分割後は全ファイルが php -l を通る。
